Tested CRUD methods.

+++ CSHARP/OENIK_PROG3_2019_2_J4G2NY/Logic/ILogic.cs
@@ -186,6 +186,6 @@ namespace Webshop.Logic
         /// </summary>
         /// <param name="price">Price.</param>
         /// <returns>Equity Ratio.</returns>
-        float EquityRatioRequestJava(int price);
+        int EquityRatioRequestJava(int price);
     }
 }
